api: Some bug fixing and improvements

+++ api/database/migrations/2020_03_14_000030_create_permission_role_table.php
@@ -19,7 +19,7 @@ class CreatePermissionRoleTable extends Migration
     {
         Schema::create('permission_role', function (Blueprint $table) {
             $table->string('permission_slug');
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->timestamp('created_at', 0)->useCurrent();
             $table->timestamp('updated_at', 0)->useCurrent();
